traced_relay: refactor traced_relay to extract library target
Extract a thin wrapper interface for traced_relay which allows
it to be included in tracebox and similar binaries layered on
top of perfetto without duplicating large amounts of code.
