+++ include/perfetto/ext/traced/traced.h
@@ -20,6 +20,7 @@
 namespace perfetto {
 
 int ServiceMain(int argc, char** argv);
+int RelayServiceMain(int argc, char** argv);
 int ProbesMain(int argc, char** argv);
 int PerfettoCmdMain(int argc, char** argv);
 int TriggerPerfettoMain(int argc, char** argv);
